+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- Forward --pr and --no-pr to gg\_one so ticket publishes merge via auto-merge pull requests by default
+
 ## [5.9.0] - 2026-07-20
 
 ### Changed
@@ -492,6 +498,7 @@ localization commands again.
 - Remove prints
 - Remove gh pr create from review
 
+[Unreleased]: https://github.com/ggsuite/gg_multi/compare/5.9.0...HEAD
 [5.9.0]: https://github.com/ggsuite/gg_multi/compare/5.8.0...5.9.0
 [5.8.0]: https://github.com/ggsuite/gg_multi/compare/5.7.1...5.8.0
 [5.7.1]: https://github.com/ggsuite/gg_multi/compare/5.7.0...5.7.1